Add InstantiatingObjectParser
Introduces InstantiatingObjectParser which is similar to the
ConstructingObjectParser, but instantiates the object using the constructor
instead of a builder function.

Closes #52499
